+++ src/material/form-field/testing/BUILD.bazel
@@ -23,27 +23,13 @@ filegroup(
 )
 
 ng_test_library(
-    name = "harness_tests_lib",
-    srcs = ["shared.spec.ts"],
+    name = "unit_tests_lib",
+    srcs = glob(["**/*.spec.ts"]),
     deps = [
         ":testing",
         "//src/cdk/testing",
         "//src/cdk/testing/private",
         "//src/cdk/testing/testbed",
-        "@npm//@angular/forms",
-        "@npm//@angular/platform-browser",
-    ],
-)
-
-ng_test_library(
-    name = "unit_tests_lib",
-    srcs = glob(
-        ["**/*.spec.ts"],
-        exclude = ["shared.spec.ts"],
-    ),
-    deps = [
-        ":harness_tests_lib",
-        ":testing",
         "//src/material/autocomplete",
         "//src/material/core",
         "//src/material/datepicker",
@@ -53,6 +39,8 @@ ng_test_library(
         "//src/material/input/testing",
         "//src/material/select",
         "//src/material/select/testing",
+        "@npm//@angular/forms",
+        "@npm//@angular/platform-browser",
     ],
 )
